cleaning up the way WeblogRequest objects are passed to Models and used for initialization.
the standard is that Models which need info about the weblog they are being used on should lookup the WeblogRequest via the 'weblogRequest' init data parameter. If they need a subclass of WeblogRequest then it is the Models job to cast the weblogRequest into the expected subclass and verify that they have what they need.



git-svn-id: https://svn.apache.org/repos/asf/incubator/roller/branches/roller_3.0@440765 13f79535-47bb-0310-9956-ffa450edef68