Some refactoring of the object model regarding Templates, Resources, and Themes.
This commit introduces a more natural OO design with regards to weblog Themes and their relationship to Templates and Resources.  Here is an overview of what these changes are ...

Template (interface) - more abstract version of the old Template interface
ThemeTemplate (interface) - effectively the same as the old Template interface, a Template tied to a Theme

Resource (interface) - a generic interface representing a static resource
ThemeResource (interface) - a Resource tied to a Theme

Theme (interface) - access to the things which compose a weblog theme
WeblogTheme (abstract class) - an instance of a Theme for a given weblog

There are then included a variety of implementation classes and subclasses for these various interfaces.



git-svn-id: https://svn.apache.org/repos/asf/incubator/roller/trunk@525922 13f79535-47bb-0310-9956-ffa450edef68