provide a fix for a formatting issue that was occuring when using the s:radio and s:checkboxlist struts2 tags.  both of those tags render their outputted items without providing a way to add a line break between each item, so they were always rendered in a horizontal list.
the fix here is to create a custom struts2 theme for roller which overrides the way those 2 tags work and add a <br/> at the end of each printed item.  this custom theme can be used only on the specific uses of that tag which need them, so the default functionality is still available.



git-svn-id: https://svn.apache.org/repos/asf/roller/trunk@541135 13f79535-47bb-0310-9956-ffa450edef68