Fixed: Variables scoping in ajax-user.js (ROL-2157) Pattern is identified and reported at  sonacloud.io as Blocker
JavaScript variable scope can be particularly difficult to understand and get right. The situation gets even worse when you consider the accidental creation of global variables, which is what happens when you declare a variable inside a function or the for clause of a for-loop without using the let, const or var keywords.